+++ python/dgl/graphbolt/impl/__init__.py
@@ -2,7 +2,6 @@
 from .basic_feature_store import *
 from .csc_sampling_graph import *
 from .neighbor_sampler import *
-from .layer_neighbor_sampler import *
 from .ondisk_dataset import *
 from .ondisk_metadata import *
 from .sampled_subgraph_impl import *
